requirements.txt
Pillow==9.2.0
tk==0.1.0
virtualenv==20.24.7
requests==2.31.0